fix IE11 breakage
